hardcode gem_path and path
shelling out to gem env path is taking too long.
It is also not possible to convert this into a properties file

These files are only used for running manageiq. so locking down the values should not be a problem.
If we go the route of installing rpms on other people's appliances, then we will need to reevaluate