use standard default file semantics for cable.yml
We currently just overwrite the `cable.yml` on every install.
This is probably fine since it is not really configurable by the end user.

But rpms typically mark a changeable configuration file as `config(noreplace)`
If the file does change, rpm creates an `.rpmsave` file to help the admin decide how to handle this conflict.

Also, this file will now be in the rpm database, so the user can tell if they modified anything.