list all vmdb directories in the gem
Overview
========

Fixes #332

The goal is to remove a warning:

```
warning: File listed twice: /var/www/miq/vmdb/certs
(a few hundred files do this)
```

Before
======

```
%defattr(-,root,root,-)
%{app_root}
%attr(-,manageiq,manageiq) %{app_root}/certs
```

We stated that we included all files recursively under app_root (/var/www/miq/vmdb).
Then we stated a few directories (with different privileges) under app_root.
So we ended up with duplicates.


Why?
====

Most app_root subdirectories are owned by root, which is the default, so using the app_root line works.

There are a few config directories that need to be editable and owned by the end user. That is why they are listed with the `%attr()` syntax.

After
=====

We explicitly listed each folder in the rails root.

I would have liked to keep on using the app_root rule so any additional files would be auto added.

This is tricky because we tend to not debug the directory listing of the `rpm` So there is a good chance that we will add a directory to app_root (vmdb / manageiq-core root) and will think it is all set but will fail on the appliance.