create an alias on upgrade
Use case
====

The product has been installed, probably enabled,
and we are upgrading the rpm (installing a later version)

`systemctl enable evmserverd` sets up alias for manageiq in the service definition.
systemctl implements this using a symbolic link.

If the alias is added in the service definition, the rpm will create the
service file, but since we are not calling systemctl enable again, nothing
will create the symbolic link.
The rpm upgrade section will need to add the missing symbolic link.

This code assumes that the service is enabled.
If it is not enabled, then the symbolic link should not be there.
Luckily, the rpm tool has quite a few PRs / tests to handle an unnecessary symbolic link.
So having the symbolic link there does not cause issues.

In the future, we will rename evmserverd to manageiq.
This code also handles that case.
It is easier to handle that use case here than having to remember how to tweak this code
when we make that change in the future.