Temporarily lock down to ruby 3.1.2-142
3.1.4-143 is now available in ubi-9-appstream.  That version of
ruby-default-gems includes time-0.2.2 and cgi that match our current gemset
gemfile.lock.  When we bundle install with 3.1.4, time and cgi are not
installed in the gemset location since they are already installed in the
system path.  If we try to use a gemset that was built with 3.1.4 on an
appliance that only has 3.1.2 (since appliances don't have ubi-9-appstream)
we fail to load time and cgi.

[root@localhost ~]# ap
/opt/...-gemset/gems/bundler-2.5.7/lib/bundler/definition.rb:596:in `materialize': Could not find cgi-0.3.6, time-0.2.2 in locally installed gems (Bundler::GemNotFound)
  from /opt/...-gemset/gems/bundler-2.5.7/lib/bundler/definition.rb:203:in `specs'
  from /opt/...-gemset/gems/bundler-2.5.7/lib/bundler/definition.rb:270:in `specs_for'
  from /opt/...-gemset/gems/bundler-2.5.7/lib/bundler/runtime.rb:18:in `setup'
  from /opt/...-gemset/gems/bundler-2.5.7/lib/bundler.rb:162:in `setup'
  from /opt/...-gemset/gems/manageiq-appliance_console-9.0.1/bin/appliance_console:6:in `<top (required)>'
  from /opt/...-gemset/bin/appliance_console:25:in `load'
  from /opt/...-gemset/bin/appliance_console:25:in `<main>'