Bring back syslog lines that were incorrectly disabled
We commented out imjournal via sed when installing the appliance.
Unfortunately, instead of commenting out just the imjournal,
we commented out to the end of the file.

This script removes all those extra comments to the end of the file.

There is a chance that we will incorrectly remove comments.
This will happen if the user has ##module(load="imklog")
If support general appliances, then we may have to factor this unlikely scenario