Move paramiko into Dockerfile.
Before
======

Installs python paramiko package via rpm
Also installs paramiko python pip package.

After
=====

Ensure paramiko is available in the environment that builds the virtual env.
virtual env building process will detect the package is available and and not try and install the paramiko pip package.
So we end up with only 1 paramiko package installed.
This will ensure we do not install the paramiko python pip package into the venv.

History
=======

This issue was introduced in
- https://github.com/ManageIQ/manageiq-rpm_build/pull/509
- 35401105fb1c8535