Exit the user entrypoint on non-zero exit code
Previously, when bundle failed following #528, the script continued
and finished the build.  It's unclear if the psych 5 error mattered
because our build still finished but we probably want to fail builds
and fix issues if the entrypoint script encounters non-zero exit codes.

Pull #546 fixed the psych 5 issue and describes the issue of the silent
error.